requirements.txt
faiss-cpu
langchain
langchain_text_splitters
langchain_community 
langchain_core 
langchain_openai
langchain_huggingface
fastapi
uvicorn
beautifulsoup4
python-dotenv
python-docx
